HTML: refactoring module for better performance.
More speed to God speed!

API breaking changes (the main API is not changed, rather internal).
Unfortunately I cannot declare functions as deprecated. A young project,
it will greatly slow down.

The code is much cleaner and the input buffers are gone.

Before that, data processing took place on the side of the tree construction.
This was done for the subsequent implementation of html parsing with
thread support.

This idea had to be abandoned. Since the current implementation is equal
in performance to the threaded one.

Now all data processing takes place in the tokenizer and the tokens already
have "correct" data.

Removed:
    - lxb_html_in_make()
    - lxb_html_in_tag_id()
    - lxb_html_in_ncasecmp()
    - lxb_html_parser_set_without_copy()
    - lxb_html_parser_set_without_copy_noi()
    - lxb_html_parser_char_process()
    - lxb_html_parser_char_copy()
    - lxb_html_parser_char_data()
    - lxb_html_parser_char_data_lcase()
    - lxb_html_parser_char_ref_data()
    - lxb_html_str_append()
    - lxb_html_token_data_calc_length()
    - lxb_html_token_make_data()
    - lxb_html_token_make_data_strict()
    - lxb_html_token_parse_data()
    - lxb_html_token_tag_id_from_data()
    - lxb_html_token_process_data()
    - lxb_html_token_attr_make_name()
    - lxb_html_token_attr_make_value()
    - lxb_html_token_attr_parse()
    - lxb_html_tokenizer_opt_set()
    - lxb_html_tokenizer_opt()
    - lxb_html_tokenizer_opt_set_noi()
    - lxb_html_tokenizer_opt_noi()

Chnaged:
    - lxb_html_token_doctype_parse()
    - lxb_html_token_find_attr()

Append:
    - lxb_html_tokenizer_attrs_make()
    - lxb_html_tokenizer_attrs_destroy()
    - lxb_html_tokenizer_attrs_set()
    - lxb_html_tokenizer_attrs()
    - lxb_html_tokenizer_attrs_mraw_set()
    - lxb_html_tokenizer_attrs_mraw()