HTML: added events (insert, remove, destroy) for elements.
For the correct application of styles to the element,
reactions to events have been added.

When an element is added to the tree, the styles for the inserted element
will be recalculated based on its Style attribute and StyleSheets applied
to the document.

Everything is the same when remove/destroy or changing an element.