Grammar: formation of the expected result (sorted).
Issue:
The result of CSS parsing can be either deterministic or arbitrary
(as the data came in and will be shown).

For example, border-top can come in many variants, but the result will
always be the same.

Input:
inset #aaa 0rem
0rem inset #aaa

Result:
0rem inset #aaa
0rem inset #aaa
0rem inset #aaa

That is, we cannot use the query to compare the result of parsing CSS
properties. We need to generate not only the query but also the expected result.

The new option:
* Forms the query and the expected result.
* The speed of forming the result has increased significantly.

New modifier:

^SORT: exclude the formation of the expected result. Query and result
will be equal.

<something> = A || B || C
<color> = <something>^SORT

New attribute:

nows: do not use spaces when merging by modifiers.

<something> = ABC
<color> = <something nows>{3, 3}

Result:
ABCABCABC

Result without `nows`:
ABC ABC ABC