Fixed object overrun during HTML serialization.
No namespace validation for the <template> element (HTMLTemplateElement)
during serialization. Only the name was checked.

<template> element which is not an HTMLTemplateElement can be serialized as
an HTMLTemplateElement, which leads to segmentation fault.

For example:
<svg><template>

In the example above, the <templete> element is not an HTMLTemplateElement.
The <template> element is in the SVG namespace.

This related rushter/selectolax#91 issue on GitHub.