CSS: fixed use-after-poison for declarations.
After each declaration is parsed, the offsets (beginning of name, end of name,
beginning of value and so on) are zeroed.
For !important offset was not zeroed, so there could be a value from the
previous declaration, if the current declaration does not have !important.

This caused the size of the declaration part to be calculated incorrectly.