Reduce the size of lexbor-encoding.so
Instead of using a pointer for the name field, embed the data in the
struct. The largest name is 4 bytes long.
On 32-bit the struct size will remain 12 bytes, but on 64-bit
the size shrinks from 16 to 12 bytes because it's less than the pointer
size. We also have to do fewer relocations.
Furthermore, by placing the data in the struct itself,
data cache locality might be better.

Results:
   text	   data	    bss	    dec	    hex	filename
4665281	1439240	      8	6104529	 5d25d1	before-patch.so
3738489	  10312	      8	3748809	 3933c9	after-patch.so