Shrink size of lexbor_shs_hash_t
lexbor_shs_hash_t is used a lot in encodings. Shrinking that struct is
good for reducing the encodings library size and improving cache
locality.
The next field has type size_t, but the key is only uint32_t. It seems,
because the hash key is only 32-bit, that next also cannot be larger
than 32-bit.

Results:
   text	   data	    bss	    dec	    hex	filename
3738489	  10312	      8	3748809	 3933c9	before-patch.so
2882009	  10312	      8	2892329	 2c2229	after-patch.so