CMake: build without C++ compiler by default.
One test required a C++ compiler.  The C++ compiler was enabled by default,
even if the project was built without tests.

The patch changes the behavior when LEXBOR_BUILD_TESTS=ON is enabled.
Now LEXBOR_BUILD_TESTS_CPP defaults to OFF, to enable C++ tests we must
explicitly specify this.

If LEXBOR_BUILD_TESTS_CPP=OFF, the C++ compiler is not enabled.

This related #197 issue on GitHub.