HTML: fixed pointer position in determine encoding function.
Previously, if there were spaces between the attribute name and the assignment
sign, the attribute value was calculated incorrectly.

The error was in the lxb_html_get_attribute() function.  This was due to the
lack of incrementation on the input buffer after checking for the assignment
sign.

For example:
<meta charset ="utf-8">

The result was:
Name: charset
Value: ="utf-8"

This closes #201 issue on GitHub.