CSS: fixed parse '|=' attribute value matching for selectors.
Selectors may contain namespace specified via '|' and an attribute value match
type with '|='.

For example:
[xml|lang="en-en"]
[lang|="en"]

In the previous, we always assumed that namespace comes before '|' and
the attribute name immediately after.

This behavior was wrong. We may have a type of attribute value match like '|='.

This closes #207 issue on GitHub.