Style: fixed NULL pointer dereferencing for style without declarations.
The style object "lxb_css_rule_style_t" must contain a Selectors and optionally
style declarations.

Previously, it was thought that a list of declarations was mandatory because of
which a null pointer was dereferenced.

The patch does a check for the existence of a list of declarations.

This closes #220, #222 issue on GitHub.