Improve performance of certain tokenizer states using SWAR
While profiling I noticed that the attribute_value_double_quoted state
and data state were very hot in the profile. This patch implements a
generic SWAR routine to seek towards an occurrence of any 4 characters
in steps of 4/8 bytes at once (depending on the bit width of the CPU).

I benchmarked this on an i7-4790 using this script:
https://gist.github.com/nielsdos/fbbcfb404fdee92e6c814212de81d015
A limited benchmark shows the following:

|                            | baseline            | double_quotes       | double_quotes + state_data |
|----------------------------|---------------------|---------------------|----------------------------|
| StackOverflow homepage     | 329.6 ms ±   3.4 ms | 288.2 ms ±   3.5 ms | 281.0 ms ±   4.6 ms        |
| English Wikipedia homepage | 289.0 ms ±   2.5 ms | 283.5 ms ±   4.5 ms | 275.9 ms ±   4.3 ms        |
| GitHub "For You" page      | 623.2 ms ±   8.0 ms | 570.1 ms ±   4.1 ms | 559.2 ms ±   4.9 ms        |