CSS: fixed use-of-uninitialized-value in parsing function.
For parsing CSS tokens, there are rules that are added/removed for each stage.
There is a first rule (parser->rules[0]) that is used at the end of the
CSS token parsing, this rule serves solely to exit the parsing.

In this rule, only one field was populated for the parsing exit phase
(parser->rules[0].phase).  But when we call the function from the phase field
(parser->rules[0].phase(...)), we pass it the context (parser->rules[0].context)
that we have uninitialized in our first rule.