Encoding: fixed incorrect code point check for ISO 2022 JP.
The encode (single and buffer) functions for the ISO 2022 JP encoding had an
incorrect codepoint check.  In code:

if (cp < 0x80) {
    switch (cp) {
        /* ... */
         case 0x00A5:
            /* ... */
         case 0x203E:
            /* ... */

From which it is clear that conditions 0x00A5 and 0x203E will never be
processed.  The patch fixes this problem and brings the comparison in line with
the specification.

Per report from Niels Dossche.

This related #239 issue on GitHub.