URL: fixed memory leak if a URL contained \n\r\t.
Previously, if a URL contained \n\r\t, these values were removed from the URL,
and a new buffer was created where valid data was placed.  Then the incoming
data stream was replaced by the created buffer with a valid URL
(without \n\r\t).

Because of replacing the incoming data stream with the created buffer, an
incorrect check when exiting the parsing function occurred.  Because of which
the created buffer was not destroyed.