HTML: fixed error report for tokens in foreign content.
For tokens in foreign content, there is no need to report an error message if
the token has a self-closing flag.

In the specification:
This error occurs if the parser encounters a start tag for an element that is
not in the list of void elements or is not a part of foreign content (i.e., not
an SVG or MathML element) that has a U+002F (/) code point right before the
closing U+003E (>) code point. The parser behaves as if the U+002F (/) is
not present.

And:
The trailing U+002F (/) in a start tag name can be used only in foreign content
to specify self-closing tags. (Self-closing tags don't exist in HTML.) It is
also allowed for void elements, but doesn't have any effect in this case.

Per report from Pavel Djundik.

This related php/php-src#17485 issue on GitHub.