CSS::Syntax: added UNICODE-RANGE token.
According to the updated specification, a new token has been added to the CSS
tokenizer <unicode-range-token>.
By default, it will not be generated.  To enable it you need to explicitly
specify it, there is a function lxb_css_syntax_tokenizer_with_unicode_range()
for this purpose.

This is a requirement of the specification:
If unicode ranges allowed is true and the input stream would start
a unicode-range, reconsume the current input code point, consume a unicode-range
token, and return it.