HTML: fixed attribute cloning for bad HTML.
Restoring bad HTML triggers the adoption agency algorithm which clones some
nodes. The problem is that when cloning nodes, the attributes of these nodes
were not cloned, but simply assigned from the original.

This could lead to memory problems if an attribute is removed (destroyed) from
a cloned element.

For example:
<a target="x"><div></a>

After parsing:
<a target="x"></a><div><a target="x"></a>

Previously, target="x" indicated the same memory in different elements.
A new attribute was created, but due to a typo in the code, the existing
attribute was assigned instead of the created one.

Per report from @nigolaj on GitHub.

This closes #273 issue on GitHub.