Selectors: code refactoring.
The :has() selector is not working correctly.
After a detailed study of the problem it became clear that the current
implementation is incorrect, and to solve the error I had to rewrite
the code for searching nodes by selectors.

Per report from Philipp Claßen (@philipp-classen) on GitHub.

This related #279 issue on GitHub.