URL: fixed "heap-buffer-overflow" for scheme is "file".
When attempting to change the hostname in a URL with the “file” scheme using
the lxb_url_api_hostname_set() function, a crash occurs.
This is caused by subtracting -1 from the beginning of the pointer, which is
an incorrect interpretation of the specification, which states:
"If state override is given and url’s scheme is "file", then decrease pointer
by 1 and set state to file host state."

In fact, we did not increase the pointer (our own implementation, our own
approach), but we left the pointer decrease, which is a mistake.

Per report from Tim Düsterhus (@TimWolla).

This related #321 issue on GitHub.